refactor: create common-deployment module and move util classes there
