+++ common-deployment/src/main/java/io/quarkiverse/operatorsdk/common/ConfigurationUtils.java
