+++ common-deployment/src/main/java/io/quarkiverse/operatorsdk/common/Constants.java
