+++ common-deployment/src/test/java/io/quarkiverse/operatorsdk/common/ConfigurationUtilsTest.java
