+++ core/deployment/pom.xml
@@ -19,7 +19,7 @@
     </dependency>
     <dependency>
       <groupId>io.quarkiverse.operatorsdk</groupId>
-      <artifactId>quarkus-operator-sdk-common</artifactId>
+      <artifactId>quarkus-operator-sdk-common-deployment</artifactId>
     </dependency>
     <dependency>
       <groupId>io.quarkus</groupId>
